requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
sqladmin==0.22.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-magic==0.4.27
psutil==5.9.6
slowapi
python-magic
python-dotenv
sqladmin
psycopg2-binary
